Modify varOne-node project package name
