+++ varOne-node/src/main/java/com/varone/node/MetricTuple.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.haredb.sparkmonitor.node;
+package com.varone.node;
 
 import java.util.List;
