+++ varOne-node/src/main/java/com/varone/node/MetricsType.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.haredb.sparkmonitor.node;
+package com.varone.node;
 
 /**
  * @author allen
