+++ varOne-node/src/main/java/com/varone/node/utils/Consts.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.haredb.sparkmonitor.node.utils;
+package com.varone.node.utils;
 
 /**
  * @author allen
