+++ varOne-web/src/main/java/com/haredb/sparkmonitor/reader/metrics/impl/MetricsRpcReaderImpl.java
@@ -15,9 +15,9 @@ import com.haredb.sparkmonitor.hadoop.rpc.protos.MetricsProtos.MetricsResponsePr
 import com.haredb.sparkmonitor.metrics.bean.MetricBean;
 import com.haredb.sparkmonitor.metrics.bean.NodeBean;
 import com.haredb.sparkmonitor.metrics.bean.TimeValuePairBean;
-import com.haredb.sparkmonitor.node.utils.MetricsDataTransfer;
 import com.haredb.sparkmonitor.reader.metrics.MetricsReader;
 import com.haredb.sparkmonitor.reader.metrics.stub.MetricsStubCenter;
+import com.varone.node.utils.MetricsDataTransfer;
 
 /**
  * @author allen
